+++ Changelog.md
@@ -7,6 +7,8 @@ This file contains the changelog of SwiftExtras.
   A view modifier that listens for a notification and performs an action when the notification is received.
 - Added [`View.onChange(of:after:perform:)`](https://0xwdg.github.io/SwiftExtras/documentation/swiftextras/swiftuicore/view/onchange(of:after:perform:))  
   A view modifier that listens for changes to a value and performs an action after the change.
+- Added [`View.onFirstAppear(perform:)`](https://0xwdg.github.io/SwiftExtras/documentation/swiftextras/swiftuicore/view/onfirstappear(perform:))  
+  A view modifier that performs an action the first time the view appears.
 
 ### 1.2.6
 - Added [`.accentColor`](https://0xwdg.github.io/SwiftExtras/documentation/swiftextras/swiftui/buttonstyle/accentcolor) to [`ButtonStyle`](https://0xwdg.github.io/SwiftExtras/documentation/swiftextras/swiftui/buttonstyle)